refactors recipes pages
