+++ src/pages/RecipeShow.js
@@ -1,0 +1,8 @@
+function RecipeShow (props) {
+    return (
+        <h1>this is the recipe show page</h1>
+    );
+}
+
+
+export default RecipeShow;
